requirements.txt
beautifulsoup4==4.3.2
cssselect==0.9.1
feedfinder2==0.0.1
feedparser==5.1.3
jieba==0.35
lxml==3.3.5
newspaper==0.0.9.8
nltk==3.2.1
numpy==1.11.3
Pillow==2.5.1
python-dateutil==2.4.0
PyYAML==3.11
requests==2.3.0
six==1.7.3
textblob==0.11.1
tldextract==1.5.1
